Move singleton to a separate header
